pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hodor"
version = "0.1.0"
description = "AI-powered code review agent that finds bugs, security issues, and logic errors in pull requests"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Karan", email = "karan@example.com"}
]
keywords = ["ai", "code-review", "github", "pr", "automation", "agent", "hodor"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # LLM & Agent Framework (OpenHands SDK)
    "openhands-sdk @ git+https://github.com/OpenHands/agent-sdk.git@main#subdirectory=openhands-sdk",
    "openhands-tools @ git+https://github.com/OpenHands/agent-sdk.git@main#subdirectory=openhands-tools",

    # Git operations (keeping gitpython for local git operations if needed)
    "gitpython>=3.1.0",

    # Code Analysis
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "radon>=6.0.0",
    "lizard>=1.17.0",

    # CLI & Output
    "rich>=13.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "ipython>=8.0.0",
]

[project.scripts]
hodor = "hodor.cli:main"

[project.urls]
Homepage = "https://github.com/mr-karan/hodor"
Documentation = "https://github.com/mr-karan/hodor/blob/main/README.md"
Repository = "https://github.com/mr-karan/hodor"
Issues = "https://github.com/mr-karan/hodor/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["hodor*"]

[tool.black]
line-length = 120
target-version = ['py313']

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=hodor --cov-report=term-missing"
